+++ PaymentService/src/main/java/CS4337/Project/PaymentService.java
@@ -1,5 +1,8 @@
 package CS4337.Project;
 
+import CS4337.Project.Shared.DTO.TransactionRequest;
+import CS4337.Project.Shared.Models.ShopItem;
+import CS4337.Project.Shared.Models.Transaction;
 import java.util.Date;
 import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -85,95 +88,5 @@ public class PaymentService {
     ResponseEntity<Map> response =
         restTemplate.exchange(
             SHOP_SERVICE_URL + "shopItem/" + shopItemId, HttpMethod.PUT, requestEntity, Map.class);
-  }
-}
-
-// TODO DRY this
-class ShopItem {
-  private int id;
-  private int shopid;
-  private double price = -1;
-  private String itemName;
-  private int stock = -1;
-  private String description;
-  private String picture;
-  private boolean isHidden;
-  private boolean canUpdate;
-
-  public int getId() {
-    return id;
-  }
-
-  public void setId(int id) {
-    this.id = id;
-  }
-
-  public int getShopid() {
-    return shopid;
-  }
-
-  public void setShopid(int shopid) {
-    this.shopid = shopid;
-  }
-
-  public double getPrice() {
-    return price;
-  }
-
-  public void setPrice(double price) {
-    this.price = price;
-  }
-
-  public String getItemName() {
-    return itemName;
-  }
-
-  public void setItemName(String itemName) {
-    this.itemName = itemName;
-  }
-
-  public int getStock() {
-    return stock;
-  }
-
-  public void setStock(int stock) {
-    this.stock = stock;
-  }
-
-  public String getDescription() {
-    return description;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
-  }
-
-  public String getPicture() {
-    return picture;
-  }
-
-  public void setPicture(String picture) {
-    this.picture = picture;
-  }
-
-  public boolean isHidden() {
-    return isHidden;
-  }
-
-  public void setHidden(boolean hidden) {
-    isHidden = hidden;
-  }
-
-  public boolean isCanUpdate() {
-    return canUpdate;
-  }
-
-  public void setCanUpdate(boolean canUpdate) {
-    this.canUpdate = canUpdate;
-  }
-
-  // Method for easily updating stock
-  public void updateStock(int amount) {
-    setStock(getStock() - amount);
   }
 }
