+++ PaymentService/src/main/java/CS4337/Project/TransactionRepository.java
@@ -1,5 +1,6 @@
 package CS4337.Project;
 
+import CS4337.Project.Shared.Models.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TransactionRepository extends CrudRepository<Transaction, Integer> {}
