+++ SharedObjs/build.gradle
@@ -28,7 +28,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.2'
     implementation 'org.projectlombok:lombok:1.18.34'
     annotationProcessor 'org.projectlombok:lombok:1.18.34'
-
+    implementation "jakarta.persistence:jakarta.persistence-api:3.1.0"
 }
 
 tasks.named('test') {
