+++ SharedObjs/src/main/java/CS4337/Project/Shared/Models/Transaction.java
@@ -1,4 +1,4 @@
-package CS4337.Project.Shared.DTO;
+package CS4337.Project.Shared.Models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -18,6 +18,7 @@ public class Transaction {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private int id;
+
   private int sourceUserId;
   private double amount;
   private String transactionStatus;
